third_party/WebKit/LayoutTests/http/tests/eventsource/eventsource-cors-redirect-expected.txt

CONSOLE ERROR: Failed to load http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Test that basic EventSource cross-origin requests fail on redirect.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS es = new EventSource("/resources/redirect.php?code=307&url=http://127.0.0.1:8080/eventsource/resources/es-cors-basic.php") did not throw exception.
PASS es.readyState is EventSource.CLOSED
PASS successfullyParsed is true

TEST COMPLETE
